+++ components/d2l-activity-editor/d2l-activity-rubrics/d2l-activity-rubrics-list-editor.js
