+++ components/d2l-activity-editor/d2l-activity-rubrics/lang/en.js
@@ -1,0 +1,5 @@
+/* eslint quotes: 0 */
+
+export default {
+	"hdrRubrics": "Rubrics" //Header for the rubrics section
+};
